fix: add trimming configuration parameter to FETCH_SRA_METADATA function (#5)
This pull request includes a change to the
`subworkflows/local/utils_nfcore_flusra_pipeline/main.nf` file to
enhance the functionality of the `FETCH_SRA_METADATA` process. The
change adds an additional parameter to the process call to include
trimming configuration.

Enhancements to the pipeline initialization:

*
[`subworkflows/local/utils_nfcore_flusra_pipeline/main.nf`](diffhunk://#diff-9e9b11abd077d43362158ffcb7149b105d61a6c1bc734a9a31900a0d5702d4b6L28-R28):
Added `params.trimming_config` as an additional parameter to the
`FETCH_SRA_METADATA` process call.