fix: update condition to use samples_to_process in workflow (#6)
This pull request includes a change to the `workflow {` block in the
`main.nf` file. The modification updates the conditional check to use
`samples_to_process` instead of `sra_accessions` from the
`PIPELINE_INITIALISATION.out`.

*
[`main.nf`](diffhunk://#diff-6401496ba455b9488ffa902a6e4d7732b2c60ff2d77c5c3ef96b28a7ac7d3b28L15-R15):
Updated the conditional check to use `samples_to_process` instead of
`sra_accessions` in the `workflow {` block.