fix: update FETCH_SRA_METADATA output to split TSV instead of CSV (#9)
This pull request includes a small but important change to the
`subworkflows/local/utils_nfcore_flusra_pipeline/main.nf` file. The
change modifies the `splitCsv` method to use a tab (`\t`) separator
instead of the default comma separator. This ensures that the CSV
parsing correctly handles tab-separated values, which may be necessary
for the specific data format being processed.

*
[`subworkflows/local/utils_nfcore_flusra_pipeline/main.nf`](diffhunk://#diff-9e9b11abd077d43362158ffcb7149b105d61a6c1bc734a9a31900a0d5702d4b6L30-R30):
Changed the `splitCsv` method to use a tab separator for parsing CSV
files.