feat: add demix_lineage_hierarchy parameter to FREYJA_DEMIX process and update Milk_Freyja workflow (#13)
This pull request includes several changes to the `FREYJA_DEMIX` process
and related configurations in the Nextflow pipeline. The changes
primarily focus on adding support for a new parameter,
`demix_lineage_hierarchy`.

### Key Changes:

#### Addition of `demix_lineage_hierarchy` parameter:

*
[`modules/nf-core/freyja/demix/main.nf`](diffhunk://#diff-a67d78be73accdfeeb6cf4083f9414bac3a2a870e5d6cd9e5ff69cd724d7d996R12):
Added `path demix_lineage_hierarchy` to the input parameters and
included the corresponding command line argument in the script section.
[[1]](diffhunk://#diff-a67d78be73accdfeeb6cf4083f9414bac3a2a870e5d6cd9e5ff69cd724d7d996R12)
[[2]](diffhunk://#diff-a67d78be73accdfeeb6cf4083f9414bac3a2a870e5d6cd9e5ff69cd724d7d996R21-R27)
*
[`nextflow.config`](diffhunk://#diff-ce7465a8e67b3b9c95696db1146ca7e6328f075e08a3d64062b87aa8608fc4eaR68-R70):
Added a new configuration parameter `demix_lineage_hierarchy` with a
default value of `null` and a note indicating it should be provided if
`demix_autoadapt` is true.
*
[`subworkflows/local/milk_freyja/main.nf`](diffhunk://#diff-9c755004458e50bc732b3102d12f4bcd28de6c07c34e40ef179e023bda01ecbeL28-R34):
Updated the `FREYJA_DEMIX` call to include the new
`demix_lineage_hierarchy` parameter.